+++ daemon/cluster/executor/backend.go
@@ -8,7 +8,6 @@ import (
 	"github.com/docker/distribution"
 	"github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/events"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/network"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/moby/moby/api/types/swarm"
@@ -16,6 +15,7 @@ import (
 	"github.com/moby/moby/api/types/volume"
 	clustertypes "github.com/moby/moby/v2/daemon/cluster/provider"
 	containerpkg "github.com/moby/moby/v2/daemon/container"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/internal/image"
 	"github.com/moby/moby/v2/daemon/libnetwork"
 	"github.com/moby/moby/v2/daemon/libnetwork/cluster"
