+++ daemon/cluster/filters.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/moby/moby/api/types/filters"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	swarmapi "github.com/moby/swarmkit/v2/api"
 )
