+++ daemon/config/builder_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/moby/moby/api/types/filters"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/fs"
 )
