+++ daemon/internal/builder-next/controller.go
@@ -45,7 +45,6 @@ import (
 	"github.com/moby/buildkit/worker"
 	"github.com/moby/buildkit/worker/containerd"
 	"github.com/moby/buildkit/worker/label"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/v2/daemon/config"
 	"github.com/moby/moby/v2/daemon/graphdriver"
 	"github.com/moby/moby/v2/daemon/internal/builder-next/adapters/containerimage"
@@ -55,6 +54,7 @@ import (
 	"github.com/moby/moby/v2/daemon/internal/builder-next/imagerefchecker"
 	mobyworker "github.com/moby/moby/v2/daemon/internal/builder-next/worker"
 	wlabel "github.com/moby/moby/v2/daemon/internal/builder-next/worker/label"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/server/buildbackend"
 	"github.com/pkg/errors"
 	bolt "go.etcd.io/bbolt"
