+++ daemon/network/filter.go
@@ -3,7 +3,7 @@ package network
 import (
 	"time"
 
-	"github.com/moby/moby/api/types/filters"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/internal/timestamp"
 	"github.com/moby/moby/v2/errdefs"
 	"github.com/pkg/errors"
