+++ daemon/network/filter_test.go
@@ -9,8 +9,8 @@ import (
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
 
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/network"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 )
 
 var _ FilterNetwork = mockFilterNetwork{}
