+++ daemon/pkg/plugin/backend_unsupported.go
@@ -9,9 +9,9 @@ import (
 	"net/http"
 
 	"github.com/distribution/reference"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/plugin"
 	"github.com/moby/moby/api/types/registry"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/server/backend"
 )
