+++ daemon/pkg/registry/search.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/distribution/registry/client/auth"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/registry"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/pkg/errors"
 )
