+++ daemon/server/router/plugin/backend.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 
 	"github.com/distribution/reference"
-	"github.com/moby/moby/api/types/filters"
 	plugintypes "github.com/moby/moby/api/types/plugin"
 	"github.com/moby/moby/api/types/registry"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/pkg/plugin"
 	"github.com/moby/moby/v2/daemon/server/backend"
 )
