+++ daemon/server/router/plugin/plugin_routes.go
@@ -9,9 +9,9 @@ import (
 	"github.com/distribution/reference"
 	"github.com/moby/moby/api/pkg/authconfig"
 	"github.com/moby/moby/api/pkg/streamformatter"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/plugin"
 	"github.com/moby/moby/api/types/registry"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/server/backend"
 	"github.com/moby/moby/v2/daemon/server/httputils"
 	"github.com/moby/moby/v2/pkg/ioutils"
