+++ daemon/server/router/volume/volume_routes.go
@@ -8,9 +8,9 @@ import (
 
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
-	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/versions"
 	"github.com/moby/moby/api/types/volume"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/server/httputils"
 	"github.com/moby/moby/v2/daemon/server/volumebackend"
 	"github.com/moby/moby/v2/daemon/volume/service/opts"
