+++ daemon/server/volumebackend/volume.go
@@ -1,6 +1,6 @@
 package volumebackend
 
-import "github.com/moby/moby/api/types/filters"
+import "github.com/moby/moby/v2/daemon/internal/filters"
 
 type ListOptions struct {
 	Filters filters.Args
