+++ daemon/volume/service/convert.go
@@ -7,9 +7,9 @@ import (
 	"time"
 
 	"github.com/containerd/log"
-	"github.com/moby/moby/api/types/filters"
 	volumetypes "github.com/moby/moby/api/types/volume"
 	"github.com/moby/moby/v2/daemon/internal/directory"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"github.com/moby/moby/v2/daemon/volume"
 	"github.com/moby/moby/v2/errdefs"
 )
