+++ daemon/volume/service/convert_test.go
@@ -3,7 +3,7 @@ package service
 import (
 	"testing"
 
-	"github.com/moby/moby/api/types/filters"
+	"github.com/moby/moby/v2/daemon/internal/filters"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
 )
